pyproject.toml
[tool.poetry]
name = "infrahub-schema"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "infrahub_schema"}]

[tool.poetry.dependencies]
python = "^3.9, < 3.13"
infrahub = {path = "../infrahub", develop = true}
invoke = "^2.1.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
jsonschema = "^4.17.0"
PyYAML = "^6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
